move test case to where other spectrum-specific tests are happening
